requirements.txt
django>=5.1.4
djangorestframework>=3.15.2
django-cors-headers>=4.4.0
pandas>=2.2.2
openpyxl>=3.1.5
